feat(design) - реализовал вызов блокчейна, пока к одной ноде
- провайдер участников
- пример файла приложен
- метод баланс

+++ api/src/Sibintek.BeerMachine/Domain/ShoppingCart.cs
@@ -9,7 +9,7 @@ namespace Sibintek.BeerMachine.Domain
     {
         public ReadOnlyCollection<Item> Items { get; }
 
-        public decimal Total { get; }
+        public long Total { get; }
 
         public ShoppingCart(List<Item> items)
         {
@@ -27,9 +27,9 @@ namespace Sibintek.BeerMachine.Domain
 
             public int Count { get; }
 
-            public decimal Total { get; }
+            public long Total { get; }
 
-            public Item(string name, decimal price, int count)
+            public Item(string name, long price, int count)
             {
                 Name = name;
                 Price = price;
